Implementando remover um veiculo

+++ aluga-aqui/src/main/java/com/alugaaqui/aluga_aqui/controller/CarrosController.java
@@ -3,8 +3,11 @@ package com.alugaaqui.aluga_aqui.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -27,9 +30,14 @@ public class CarrosController {
        return carros.findAll();
 
     }
-    @GetMapping("/delete{id}")
-    public void deleteById(Long id) {
+   @DeleteMapping("remover/{id}")
+public ResponseEntity<Void> deleteById(@PathVariable long id) {
+    if (carros.existsById(id)) {
         carros.deleteById(id);
+        return ResponseEntity.noContent().build(); // 204 No Content
+    } else {
+        return ResponseEntity.notFound().build(); // 404 Not Found
     }
+}
     
 }
